+++ pvl/collections.py
@@ -26,17 +26,30 @@ class (formerly the Units class).
 
 import pprint
 import warnings
-from collections import namedtuple
-from collections.abc import Mapping, MutableMapping, MutableSequence
+from abc import abstractmethod
+from collections import namedtuple, abc
 
 
 class MutableMappingSequence(
-    MutableMapping, MutableSequence
+    abc.MutableMapping, abc.MutableSequence
 ):
     """ABC for a mutable object that has both mapping and
     sequence characteristics.
+
+    Must implement `.getall()` and `.popall()` since MutableMappingSequence
+    can have many values for a single key, while `.get(k)` and
+    `.pop(k)` return and operate on a single value, the *all*
+    versions return and operate on all values in the MutableMappingSequence
+    with the key `k`.
     """
-    pass
+
+    @abstractmethod
+    def getall(self, key):
+        pass
+
+    @abstractmethod
+    def popall(self, key):
+        pass
 
 
 dict_setitem = dict.__setitem__
@@ -202,9 +215,8 @@ class OrderedMultiDict(dict, MutableMappingSequence):
 
         return "{!s}([\n{!s}\n])".format(type(self).__name__, '\n'.join(lines))
 
-    get = MutableMapping.get
-    update = MutableMapping.update
-    pop = MutableMapping.pop
+    get = abc.MutableMapping.get
+    update = abc.MutableMapping.update
 
     def keys(self):
         return KeysView(self)
@@ -243,7 +255,7 @@ class OrderedMultiDict(dict, MutableMappingSequence):
 
         iterable = args[0] if args else None
         if iterable:
-            if isinstance(iterable, Mapping) or hasattr(iterable, 'items'):
+            if isinstance(iterable, abc.Mapping) or hasattr(iterable, 'items'):
                 for key, value in iterable.items():
                     self.append(key, value)
             elif hasattr(iterable, 'keys'):
@@ -256,14 +268,33 @@ class OrderedMultiDict(dict, MutableMappingSequence):
         for key, value in kwargs.items():
             self.append(key, value)
 
-    def getlist(self, key):
+    def getall(self, key) -> abc.Sequence:
+        """Returns a list of all the values for a named field.
+        Returns KeyError if the key doesn't exist.
+        """
+        return list(dict_getitem(self, key))
+
+    def getlist(self, key) -> abc.Sequence:
         """Returns a list of all the values for the named field.
         Returns an empty list if the key doesn't exist.
         """
+        warnings.warn(
+            "The pvl.collections.OrderedMultiDict.getlist(k) function is "
+            "deprecated in favor of .getall(), please begin using it, as "
+            ".getlist() may be removed in the next major patch.",
+            PendingDeprecationWarning
+        )
+
         try:
-            return list(dict_getitem(self, key))
+            return self.getall(key)
         except KeyError:
             return []
+
+    # Turns out that this super-class function, even though it doesn't have
+    # a concept of multiple keys, clears out multiple elements with the key,
+    # probably because of how __delitem__ is defined:
+    popall = abc.MutableMapping.pop
+    pop = abc.MutableMapping.pop
 
     def popitem(self):
         if not self:
@@ -282,6 +313,45 @@ class OrderedMultiDict(dict, MutableMappingSequence):
     def copy(self):
         return type(self)(self)
 
+    def insert(self, index: int, *args) -> None:
+        """Inserts at the index given by *index*.
+
+        The first positional argument will be taken as the
+        *index*. If three arguments are given, the second will be taken
+        as the *key*, and the third as the *value*.  If only two arguments are
+        given, the second must be a two-element sequence, where the first will
+        be the *key* and the second the *value*.
+        """
+        if len(args) == 1:
+            if len(args[0]) == 2:
+                key, value = args[0]
+            else:
+                raise IndexError(
+                    "If a sequence is provided to the second positional "
+                    f"argument of pvl.OrderedMultiDict.insert() it must have "
+                    f"exactly 2 elements, but it is {args[0]}"
+                )
+        elif len(args) == 2:
+            self.__items.insert(index, args)
+            key, value = args
+        else:
+            raise TypeError(
+                f"{self.__name__}.insert() takes 2 or 3 positional arguments, "
+                f"but {len(args)} were given."
+            )
+
+        self.__items.insert(index, (key, value))
+
+        # Make sure indexing works with the new item
+        if key in self:
+            value_list = [val for k, val in self.__items if
+                          k == key]
+            dict_setitem(self, key, value_list)
+        else:
+            dict_setitem(self, key, [value])
+
+        return
+
     def __insert_wrapper(func):
         """Make sure the arguments given to the insert methods are correct."""
 
@@ -295,10 +365,10 @@ class OrderedMultiDict(dict, MutableMappingSequence):
             return func(self, key, new_item, instance)
         return check_func
 
-    def _get_index_for_insert(self, key, instance):
+    def _get_index_for_insert(self, key, instance: int) -> int:
         """Get the index of the key to insert before or after."""
         if instance == 0:
-            # Index method will return the first occurence of the key
+            # Index method will return the first occurrence of the key
             index = self.keys().index(key)
         else:
             occurrence = -1
@@ -306,37 +376,38 @@ class OrderedMultiDict(dict, MutableMappingSequence):
                 if k == key:
                     occurrence += 1
                     if occurrence == instance:
-                        # Found the key and the correct occurence of the key
+                        # Found the key and the correct occurrence of the key
                         break
 
             if occurrence != instance:
                 # Gone through the entire list of keys and the instance number
-                # given is too high for the number of occurences of the key
+                # given is too high for the number of occurrences of the key
                 raise ValueError(f"Cannot insert before/after the {instance} "
                                  f"instance of the key '{key}' since there are "
-                                 f"only {occurrence} occurences of the key")
+                                 f"only {occurrence} occurrences of the key")
         return index
 
-    def _insert_item(self, key, new_item, instance, is_after):
+    def _insert_item(
+            self, key, new_item: abc.Iterable, instance: int, is_after: bool
+    ):
         """Insert a new item before or after another item."""
         index = self._get_index_for_insert(key, instance)
         index = index + 1 if is_after else index
-        self.__items = self.__items[:index] + new_item + self.__items[index:]
-        # Make sure indexing works with new items
-        for new_key, new_value in new_item:
-            if new_key in self:
-                value_list = [val for k, val in self.__items if k == new_key]
-                dict_setitem(self, new_key, value_list)
-            else:
-                dict_setitem(self, new_key, [new_value])
+
+        # But new_item is always a list of two-tuples, even if only one, and
+        # all should be inserted, so despite the singular "an item"
+        # in the doc strings, this could be a whole bunch.
+        for pair in new_item:
+            self.insert(index, pair)
+            index += 1
 
     @__insert_wrapper
-    def insert_after(self, key, new_item, instance=0):
+    def insert_after(self, key, new_item: abc.Iterable, instance=0):
         """Insert an item after a key"""
         self._insert_item(key, new_item, instance, True)
 
     @__insert_wrapper
-    def insert_before(self, key, new_item, instance=0):
+    def insert_before(self, key, new_item: abc.Iterable, instance=0):
         """Insert an item before a key"""
         self._insert_item(key, new_item, instance, False)
 
@@ -376,9 +447,10 @@ class Quantity(namedtuple('Quantity', ['value', 'units'])):
     """
     pass
 
+
 class Units(Quantity):
     warnings.warn(
         "The pvl.collections.Units object is deprecated, and may be removed at "
         "the next major patch. Please use pvl.collections.Quantity instead.",
-        DeprecationWarning
+        PendingDeprecationWarning
     )